refactor: split ui setup diagnostics
